requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
accelerate=1.2.0=pyhd8ed1ab_0
aiohappyeyeballs=2.4.4=pyhd8ed1ab_1
aiohttp=3.11.10=py312h178313f_0
aiosignal=1.3.1=pyhd8ed1ab_1
alsa-lib=1.2.13=hb9d3cd8_0
annotated-types=0.7.0=pypi_0
aom=3.6.1=h59595ed_0
attrs=24.2.0=pyh71513ae_1
blas=1.1=openblas
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
brotli-python=1.1.0=py312h2ec8cdc_2
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.11.26=h06a4308_0
cairo=1.18.2=h3394656_1
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py312h06ac9bb_0
chardet=5.2.0=py312h7900ff3_2
charset-normalizer=3.4.0=pyhd8ed1ab_1
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
contourpy=1.3.1=py312h68727a3_0
cpython=3.12.8=py312hd8ed1ab_1
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.77=0
cuda-runtime=12.4.1=0
cuda-version=12.6=3
cycler=0.12.1=pyhd8ed1ab_1
cyrus-sasl=2.1.27=h54b06d7_7
dbus=1.13.18=hb2f20db_0
docker-pycreds=0.4.0=pypi_0
double-conversion=3.3.0=h59595ed_0
expat=2.6.4=h5888daf_0
ffmpeg=4.4.2=gpl_hdf48244_113
filelock=3.16.1=pyhd8ed1ab_1
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.55.2=py312h178313f_0
freetype=2.12.1=h267a509_2
freetype-py=2.3.0=pyhd8ed1ab_0
frozenlist=1.5.0=py312h66e93f0_0
fsspec=2024.10.0=pyhd8ed1ab_1
gettext=0.22.5=he02047a_3
gettext-tools=0.22.5=he02047a_3
giflib=5.2.2=hd590300_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
glib=2.82.2=h44428e9_0
glib-tools=2.82.2=h4833e2c_0
gmp=6.3.0=hac33072_2
gmpy2=2.1.5=py312h7201bc8_3
gnutls=3.7.9=hb077bed_0
graphite2=1.3.14=h295c915_1
greenlet=3.1.1=py312h2ec8cdc_0
h2=4.1.0=pyhd8ed1ab_1
harfbuzz=9.0.0=hda332d3_1
hpack=4.0.0=pyhd8ed1ab_1
huggingface_hub=0.26.5=pyhd8ed1ab_1
hyperframe=6.0.1=pyhd8ed1ab_1
icu=75.1=he02047a_0
idna=3.10=pyhd8ed1ab_1
jinja2=3.1.4=pyhd8ed1ab_1
joblib=1.4.2=pyhd8ed1ab_1
kaldi=5.5.1112=cpu_hd7b63f8_5
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py312h68727a3_0
krb5=1.21.3=h659f571_0
lame=3.100=h166bdaf_1003
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_2
lerc=4.0.0=h27087fc_0
libabseil=20240722.0=cxx17_h5888daf_1
libasprintf=0.22.5=he8f35ee_3
libasprintf-devel=0.22.5=he8f35ee_3
libblas=3.9.0=25_linux64_openblas
libboost=1.86.0=h6c02f8c_3
libboost-python=1.86.0=py312hc39e661_3
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=25_linux64_openblas
libclang-cpp19.1=19.1.5=default_hb5137d0_0
libclang13=19.1.5=default_h9c6a7e4_0
libcublas=12.4.5.8=0
libcufft=11.2.1.3=0
libcufile=1.11.1.6=0
libcups=2.3.3=h4637d8d_4
libcurand=10.3.7.77=0
libcusolver=11.6.1.9=0
libcusparse=12.3.1.170=0
libdeflate=1.22=hb9d3cd8_0
libdrm=2.4.124=hb9d3cd8_0
libedit=3.1.20230828=h5eee18b_0
libegl=1.7.0=ha4b6fd6_2
libexpat=2.6.4=h5888daf_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgettextpo=0.22.5=he02047a_3
libgettextpo-devel=0.22.5=he02047a_3
libgfortran=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgl=1.7.0=ha4b6fd6_2
libglib=2.82.2=h2ff4ddf_0
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libgomp=14.2.0=h77fa898_1
libhwloc=2.11.2=default_h0d58e46_1001
libiconv=1.17=hd590300_2
libidn2=2.3.7=hd590300_0
libjpeg-turbo=3.0.3=h5eee18b_0
liblapack=3.9.0=25_linux64_openblas
liblapacke=3.9.0=25_linux64_openblas
libllvm19=19.1.5=ha7bfdaf_0
liblzma=5.6.3=hb9d3cd8_1
liblzma-devel=5.6.3=hb9d3cd8_1
libnpp=12.2.5.30=0
libnsl=2.0.1=hd590300_0
libntlm=1.6=h7f8727e_0
libnvfatbin=12.6.77=0
libnvjitlink=12.4.127=0
libnvjpeg=12.3.1.117=0
libopenblas=0.3.28=pthreads_h94d23a6_1
libopengl=1.7.0=ha4b6fd6_2
libpciaccess=0.18=hd590300_0
libpng=1.6.44=hadc24fc_0
libpq=17.2=h3b95a9b_1
libprotobuf=5.28.2=h5b01275_0
librdkit=2024.09.3=h84b0b3c_0
libsqlite=3.47.2=hee588c1_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtasn1=4.19.0=h166bdaf_0
libtiff=4.7.0=hc4654cb_2
libtorch=2.4.1=cpu_mkl_h654cdd8_103
libunistring=0.9.10=h7f98852_0
libuuid=2.38.1=h0b41bf4_0
libuv=1.49.2=hb9d3cd8_0
libva=2.22.0=h8a09558_1
libvpx=1.13.1=h59595ed_0
libwebp=1.4.0=h2c329e2_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.13.5=h8d12d68_1
libxslt=1.1.41=h097e994_0
libzlib=1.3.1=hb9d3cd8_2
llvm-openmp=19.1.5=h024ca30_0
markupsafe=3.0.2=py312h178313f_1
matplotlib=3.9.3=py312h7900ff3_0
matplotlib-base=3.9.3=py312hd3ec401_0
mkl=2023.2.0=h84fe81f_50496
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_1
multidict=6.1.0=py312h178313f_1
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=9.0.1=h266115a_3
mysql-libs=9.0.1=he0572af_3
ncurses=6.5=he02047a_1
nettle=3.9.1=h7ab15ed_0
networkx=3.4.2=pyh267e887_2
nomkl=3.0=0
numpy=2.2.0=py312h7e784f5_0
openblas=0.3.28=pthreads_h6ec200e_1
openfst=1.8.3=h84d6215_3
openh264=2.3.1=hcb278e6_2
openjpeg=2.5.3=h5fbd93e_0
openldap=2.6.9=he970967_0
openssl=3.4.0=hb9d3cd8_0
p11-kit=0.24.1=hc5aa10d_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py312hf9745cd_1
pcre2=10.44=hba22ea6_2
pillow=11.0.0=py312h7b63e92_0
pip=24.3.1=pyh8b19718_0
pixman=0.44.2=h29eaf8c_0
platformdirs=4.3.6=pypi_0
propcache=0.2.1=py312h66e93f0_0
protobuf=5.29.1=pypi_0
psutil=6.1.0=py312h66e93f0_0
pthread-stubs=0.4=hb9d3cd8_1002
pycairo=1.27.0=py312h51b637d_0
pycparser=2.22=pyh29332c3_1
pydantic=2.10.3=pypi_0
pydantic-core=2.27.1=pypi_0
pyg=2.6.1=py312_torch_2.4.0_cu124
pyparsing=3.2.0=pyhd8ed1ab_2
pyside6=6.8.0.2=py312h91f0f75_0
pysocks=1.7.1=pyha55dd90_7
python=3.12.8=h9e4cc4f_1_cpython
python-dateutil=2.9.0post0=py312h06a4308_2
python-tzdata=2024.2=pyhd8ed1ab_1
python_abi=3.12=5_cp312
pytorch=2.4.1=cpu_mkl_py312h1b0a35b_103
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.2=py312h66e93f0_1
qhull=2020.2=h434a139_5
qt6-main=6.8.0=h6e8976b_0
rdkit=2024.09.3=py312h07adbc5_0
readline=8.2=h8228510_1
reportlab=4.2.5=py312h66e93f0_0
requests=2.32.3=pyhd8ed1ab_1
rlpycairo=0.2.0=pyhd8ed1ab_0
safetensors=0.4.5=py312h12e396e_0
scikit-learn=1.6.0=py312h7a48858_0
scipy=1.14.1=py312h62794b6_2
sentry-sdk=2.19.2=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=75.6.0=pyhff2d567_1
six=1.17.0=pyhd8ed1ab_0
sleef=3.7=h1b44611_2
smmap=5.0.1=pypi_0
sqlalchemy=2.0.36=py312h66e93f0_0
svt-av1=1.4.1=hcb278e6_0
sympy=1.13.3=pyh2585a3b_104
tbb=2021.13.0=hceb3a55_1
threadpoolctl=3.5.0=pyhc1e730c_0
tk=8.6.13=noxft_h4845f30_101
torchaudio=2.4.1=cpu_py312hd217f58_2
torchvision=0.19.1=cpu_py312h28d33e9_1
tornado=6.4.2=py312h66e93f0_0
tqdm=4.67.1=pyhd8ed1ab_0
typing-extensions=4.12.2=hd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024b=hc8b5060_0
unicodedata2=15.1.0=py312h66e93f0_1
urllib3=2.2.3=pyhd8ed1ab_1
wandb=0.19.0=pypi_0
wayland=1.23.1=h3e06ad9_0
wayland-protocols=1.37=hd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.5=hb9d3cd8_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.43=hb9d3cd8_0
xorg-libice=1.1.1=hb9d3cd8_1
xorg-libsm=1.2.4=he73a12e_1
xorg-libx11=1.8.10=h4f16b4b_1
xorg-libxau=1.0.11=hb9d3cd8_1
xorg-libxcomposite=0.4.6=hb9d3cd8_2
xorg-libxcursor=1.2.3=hb9d3cd8_0
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxrandr=1.5.4=hb9d3cd8_0
xorg-libxrender=0.9.11=hb9d3cd8_2
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.5=hb9d3cd8_4
xz=5.6.3=hbcc6ac9_1
xz-gpl-tools=5.6.3=hbcc6ac9_1
xz-tools=5.6.3=hb9d3cd8_1
yaml=0.2.5=h7f98852_2
yarl=1.18.3=py312h66e93f0_0
zstandard=0.23.0=py312hef9b889_1
zstd=1.5.6=ha6fb4c9_0
